pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "angels-book-llm"
version = "0.1.0"
description = "Training and inference for a small decoder-only LLM on the Angels Book dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "PyYAML>=6.0",
    "rank_bm25>=0.2.2",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
